+++ app/src/main/java/com/eveningoutpost/dexdrip/ui/charts/BasalChart.java
@@ -65,7 +65,7 @@ public class BasalChart {
     }
 
     public static void refreshAxis(final ColumnChartView chart) {
-        chart.getChartData().setAxisYRight(chartYAxis(chart.getMaximumViewport().top));
+        chart.getChartData().setAxisYRight(chartYAxis((float) chart.getMaximumViewport().top));
     }
 
     private static List<Column> getColumns(final String profileName, @ColorInt int color) {
@@ -148,7 +148,7 @@ public class BasalChart {
         float max_height = 0.2f;
         for (Column col : lineData.getColumns()) {
             final SubcolumnValue sub = col.getValues().get(0);
-            final float val = Math.max(sub.getValue(), sub.getImmutableOriginValue());
+            final float val = (float) Math.max(sub.getValue(), sub.getImmutableOriginValue());
             if (val > max_height) {
                 max_height = val;
             }
